Move editor tabs from the sources reducer to a new reducer
