+++ src/selectors/index.js
@@ -6,6 +6,7 @@
 
 export * from "../reducers/expressions";
 export * from "../reducers/sources";
+export * from "../reducers/tabs";
 export * from "../reducers/pause";
 export * from "../reducers/debuggee";
 export * from "../reducers/breakpoints";
